+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Resolves Install USB Creation using incorrect installer
 - Resolves `installer` failing to extract InstallAssistant in older OSes
 - Resolves certain Samsung NVMe drives appearing as external on Mac Pros
+- Add FeatureUnlock configurability 
 
 ## 0.3.3
 - Disable Asset Caching support with spoofless approach